chore: split form away from calculation table
